Moved Plant and Garden into their own files
