+++ src/App.css
@@ -1,7 +1,3 @@
-.Garden {
-  display: block;
-}
-
 .App {
   text-align: center;
 }
