+++ src/Garden.css
@@ -1,0 +1,4 @@
+.Garden {
+  display: block;
+}
+
